Remove extremely sparse columns

remove_sparse() removes columns from sparse and dense matrices that have a sparsity value greater than a user-defined threshold.

Usage

remove_sparse(x, threshold)

# S3 method for dgCMatrix
remove_sparse(x, threshold)

# S3 method for matrix
remove_sparse(x, threshold)

Arguments

x

A matrix or dgCMatrix.

threshold

A double between 0 and 1 specifying the sparsity threshold at which to remove columns.

Value

x with sparse columns removed.

Details

remove_sparse() is an S3 generic with methods for:

dgCMatrix

matrix

Examples

# Create a sparse matrix with very sparse columns
x <- Matrix::rsparsematrix(10, 5, 0.1)
colnames(x) <- paste0("x", 1:5)
# Print x
x
#> 10 x 5 sparse Matrix of class "dgCMatrix"
#>         x1  x2  x3    x4 x5
#>  [1,] -1.2 .   .   -0.21  .
#>  [2,]  .   .   .   -0.22  .
#>  [3,]  .   .   .    .     .
#>  [4,]  .   .   .    .     .
#>  [5,]  .   .   .    .     .
#>  [6,]  .   1.6 1.2  .     .
#>  [7,]  .   .   .    .     .
#>  [8,]  .   .   .    .     .
#>  [9,]  .   .   .    .     .
#> [10,]  .   .   .    .     .

# Same matrix in dense format
xdense <- as.matrix(x)

# Drop duplicate columns
remove_sparse(x, threshold = 0.9)
#> 10 x 1 sparse Matrix of class "dgCMatrix"
#>          x4
#>  [1,] -0.21
#>  [2,] -0.22
#>  [3,]  .   
#>  [4,]  .   
#>  [5,]  .   
#>  [6,]  .   
#>  [7,]  .   
#>  [8,]  .   
#>  [9,]  .   
#> [10,]  .   
remove_sparse(xdense, threshold = 0.9)
#>          x4
#>  [1,] -0.21
#>  [2,] -0.22
#>  [3,]  0.00
#>  [4,]  0.00
#>  [5,]  0.00
#>  [6,]  0.00
#>  [7,]  0.00
#>  [8,]  0.00
#>  [9,]  0.00
#> [10,]  0.00