Transform Sparse Matrix columns

transform_cols() transforms specified matrix columns with a user-supplied function.

Usage

transform_cols(x, fns, ..., which.cols, drop, name.sep)

# S3 method for CsparseMatrix
transform_cols(x, fns, ..., which.cols, drop = FALSE, name.sep = NULL)

# S3 method for matrix
transform_cols(x, fns, ..., which.cols, drop = FALSE, name.sep = NULL)

Arguments

x

A matrix or CsparseMatrix.

fns

A user-supplied function, or list of functions, to apply to the specified columns.

...

Additional arguments to pass to fns. It is important to note that these arguments will be passed to every function in fns, and so should only include arguments that are relevant to each function.

which.cols

A numeric vector indicating column indices or a character vector indicating column names.

drop

A logical value. If the functions in fns work with columnar matrices, then set drop = FALSE, otherwise if the functions in fns require a numeric vector, set drop = TRUE. When working with large sparse matrices, it is essential to set drop = FALSE, as the alternative will be much more memory intensive.

name.sep

A NULL value or a list corresponding to each element of fns. If name.sep is NULL the specified columns will be transformed in-place, provided length(fns) == 1. If fns has more than one element, and name.sep is NULL, it will default to a numeric vector that is equal to length(fns). If name.sep is provided as a list, each element of this list must contain a character vector of length 1 or nrow(x) that will be appended to existing column names to create new column names. Providing this argument ensures that the transformed columns will be appended as new matrix columns.

Details

transform_cols() is an S3 generic with methods for:

CsparseMatrix

matrix

Examples

x <- Matrix::rsparsematrix(10, 4, .9)
colnames(x) <- paste0("x", 1:4)
x
#> 10 x 4 sparse Matrix of class "dgCMatrix"
#>           x1    x2     x3    x4
#>  [1,] -0.110 -0.18 -0.094 -0.29
#>  [2,]  1.700 -0.82  1.400  0.64
#>  [3,]  0.640  0.61 -0.910 -0.66
#>  [4,]  0.042 -0.25 -0.054  0.79
#>  [5,] -0.660 -0.43  0.057 -0.47
#>  [6,]  0.620  0.09  1.700  .   
#>  [7,] -0.920  0.47  0.180  .   
#>  [8,] -1.500  .     0.510 -0.68
#>  [9,]  0.450  1.60  0.720 -0.64
#> [10,] -1.000  .     0.130  0.74

# Convert columns in-place with a single function
transform_cols(x, fns = function(i) i^2, which.cols = 3:4)
#> 10 x 4 sparse Matrix of class "dgCMatrix"
#>           x1    x2       x3     x4
#>  [1,] -0.110 -0.18 0.008836 0.0841
#>  [2,]  1.700 -0.82 1.960000 0.4096
#>  [3,]  0.640  0.61 0.828100 0.4356
#>  [4,]  0.042 -0.25 0.002916 0.6241
#>  [5,] -0.660 -0.43 0.003249 0.2209
#>  [6,]  0.620  0.09 2.890000 .     
#>  [7,] -0.920  0.47 0.032400 .     
#>  [8,] -1.500  .    0.260100 0.4624
#>  [9,]  0.450  1.60 0.518400 0.4096
#> [10,] -1.000  .    0.016900 0.5476
transform_cols(x, fns = function(i) i^2, which.cols = c("x3", "x4"))
#> 10 x 4 sparse Matrix of class "dgCMatrix"
#>           x1    x2       x3     x4
#>  [1,] -0.110 -0.18 0.008836 0.0841
#>  [2,]  1.700 -0.82 1.960000 0.4096
#>  [3,]  0.640  0.61 0.828100 0.4356
#>  [4,]  0.042 -0.25 0.002916 0.6241
#>  [5,] -0.660 -0.43 0.003249 0.2209
#>  [6,]  0.620  0.09 2.890000 .     
#>  [7,] -0.920  0.47 0.032400 .     
#>  [8,] -1.500  .    0.260100 0.4624
#>  [9,]  0.450  1.60 0.518400 0.4096
#> [10,] -1.000  .    0.016900 0.5476

# Mutate new columns with a single function
transform_cols(x, fns = scale, which.cols = 3:4, name.sep = "scaled")
#> 10 x 6 sparse Matrix of class "dgCMatrix"
#>           x1    x2     x3    x4  x3_scaled   x4_scaled
#>  [1,] -0.110 -0.18 -0.094 -0.29 -0.6026003 -0.39313052
#>  [2,]  1.700 -0.82  1.400  0.64  1.3635165  1.17601706
#>  [3,]  0.640  0.61 -0.910 -0.66 -1.6764634 -1.01741505
#>  [4,]  0.042 -0.25 -0.054  0.79 -0.5499600  1.42910538
#>  [5,] -0.660 -0.43  0.057 -0.47 -0.4038830 -0.69683651
#>  [6,]  0.620  0.09  1.700  .     1.7583191  0.09617356
#>  [7,] -0.920  0.47  0.180  .    -0.2420140  0.09617356
#>  [8,] -1.500  .     0.510 -0.68  0.1922689 -1.05116015
#>  [9,]  0.450  1.60  0.720 -0.64  0.4686307 -0.98366994
#> [10,] -1.000  .     0.130  0.74 -0.3078144  1.34474261

# Mutate new columns with a list of functions and names
transform_cols(
  x,
  fns = c(function(i) i^2, function(i) i^3),
  which.cols = 3:4,
  name.sep = c("squared", "cubed")
)
#> 10 x 8 sparse Matrix of class "dgCMatrix"
#>           x1    x2     x3    x4 x3_squared x4_squared     x3_cubed  x4_cubed
#>  [1,] -0.110 -0.18 -0.094 -0.29   0.008836     0.0841 -0.000830584 -0.024389
#>  [2,]  1.700 -0.82  1.400  0.64   1.960000     0.4096  2.744000000  0.262144
#>  [3,]  0.640  0.61 -0.910 -0.66   0.828100     0.4356 -0.753571000 -0.287496
#>  [4,]  0.042 -0.25 -0.054  0.79   0.002916     0.6241 -0.000157464  0.493039
#>  [5,] -0.660 -0.43  0.057 -0.47   0.003249     0.2209  0.000185193 -0.103823
#>  [6,]  0.620  0.09  1.700  .      2.890000     .       4.913000000  .       
#>  [7,] -0.920  0.47  0.180  .      0.032400     .       0.005832000  .       
#>  [8,] -1.500  .     0.510 -0.68   0.260100     0.4624  0.132651000 -0.314432
#>  [9,]  0.450  1.60  0.720 -0.64   0.518400     0.4096  0.373248000 -0.262144
#> [10,] -1.000  .     0.130  0.74   0.016900     0.5476  0.002197000  0.405224

# Mutate new columns with a list of functions and names for each new column
transform_cols(
  x,
  fns = c(function(i) i^2, function(i) i^3),
  which.cols = 3:4,
  name.sep = list(paste0("squared", 1:2), paste0("cubed", 1:2))
)
#> 10 x 8 sparse Matrix of class "dgCMatrix"
#>           x1    x2     x3    x4 x3_squared1 x4_squared2    x3_cubed1 x4_cubed2
#>  [1,] -0.110 -0.18 -0.094 -0.29    0.008836      0.0841 -0.000830584 -0.024389
#>  [2,]  1.700 -0.82  1.400  0.64    1.960000      0.4096  2.744000000  0.262144
#>  [3,]  0.640  0.61 -0.910 -0.66    0.828100      0.4356 -0.753571000 -0.287496
#>  [4,]  0.042 -0.25 -0.054  0.79    0.002916      0.6241 -0.000157464  0.493039
#>  [5,] -0.660 -0.43  0.057 -0.47    0.003249      0.2209  0.000185193 -0.103823
#>  [6,]  0.620  0.09  1.700  .       2.890000      .       4.913000000  .       
#>  [7,] -0.920  0.47  0.180  .       0.032400      .       0.005832000  .       
#>  [8,] -1.500  .     0.510 -0.68    0.260100      0.4624  0.132651000 -0.314432
#>  [9,]  0.450  1.60  0.720 -0.64    0.518400      0.4096  0.373248000 -0.262144
#> [10,] -1.000  .     0.130  0.74    0.016900      0.5476  0.002197000  0.405224