Transform Sparse Matrix columns

transform_cols() transforms specified matrix columns with a user-supplied function.

Usage

transform_cols(x, fun, ..., which, drop, name.sep)

# S3 method for CsparseMatrix
transform_cols(x, fun, ..., which, drop = FALSE, name.sep = NULL)

# S3 method for matrix
transform_cols(x, fun, ..., which, drop = FALSE, name.sep = NULL)

Arguments

x

A matrix or CsparseMatrix.

fun

A user-supplied function to apply to the specified columns.

...

Additional arguments to pass to fun.

which

A numeric vector indicating column indices or a character vector indicating column names.

drop

A logical value. If fun works with columnar matrices, then set drop = FALSE, otherwise if fun requires a numeric vector set drop = TRUE.

name.sep

A character vector of length 1 or nrow(x) that will be appended to existing column names to create new columns. Providing this argument ensures that the transformed columns will be appended as new matrix columns. If name.sep is NULL the specified columns will be transformed in-place. If name.sep is provided, but is faulty, the specified columns will be appended as new columns but will have un-changed column names.

Details

transform_cols() is an S3 generic with methods for:

CsparseMatrix

matrix

Examples

x <- Matrix::rsparsematrix(10, 4, .9)
colnames(x) <- paste0("x", 1:4)
x
#> 10 x 4 sparse Matrix of class "dgCMatrix"
#>          x1     x2     x3     x4
#>  [1,] -0.11 -0.052  0.045 -0.320
#>  [2,]  0.28  .      0.790  .    
#>  [3,] -1.50  1.400  1.300 -0.075
#>  [4,]  0.43  0.560 -0.710 -0.088
#>  [5,]  1.00 -0.480  .     -0.071
#>  [6,] -0.71  2.000 -1.600  0.081
#>  [7,]  0.52 -0.800 -0.610 -1.500
#>  [8,] -1.20  1.300  0.530 -0.590
#>  [9,]  .    -0.140 -0.120 -0.160
#> [10,]  1.40  0.460 -1.100 -1.400

transform_cols(x, fun = function(i) i^2, which = paste0("x", 3:4))
#> 10 x 4 sparse Matrix of class "dgCMatrix"
#>          x1     x2       x3       x4
#>  [1,] -0.11 -0.052 0.002025 0.102400
#>  [2,]  0.28  .     0.624100 .       
#>  [3,] -1.50  1.400 1.690000 0.005625
#>  [4,]  0.43  0.560 0.504100 0.007744
#>  [5,]  1.00 -0.480 .        0.005041
#>  [6,] -0.71  2.000 2.560000 0.006561
#>  [7,]  0.52 -0.800 0.372100 2.250000
#>  [8,] -1.20  1.300 0.280900 0.348100
#>  [9,]  .    -0.140 0.014400 0.025600
#> [10,]  1.40  0.460 1.210000 1.960000
transform_cols(x, fun = scale, which = 3:4, drop = TRUE)
#> 10 x 4 sparse Matrix of class "dgCMatrix"
#>          x1     x2          x3         x4
#>  [1,] -0.11 -0.052  0.21765254  0.1595718
#>  [2,]  0.28  .      1.05999612  0.7128002
#>  [3,] -1.50  1.400  1.63663400  0.5831373
#>  [4,]  0.43  0.560 -0.63599767  0.5606624
#>  [5,]  1.00 -0.480  0.16677272  0.5900526
#>  [6,] -0.71  2.000 -1.64228732  0.8528361
#>  [7,]  0.52 -0.800 -0.52293142 -1.8804579
#>  [8,] -1.20  1.300  0.76602386 -0.3072146
#>  [9,]  .    -0.140  0.03109322  0.4361860
#> [10,]  1.40  0.460 -1.07695605 -1.7075740
transform_cols(x, fun = scale, which = 3:4, name.sep = "normalized")
#> 10 x 6 sparse Matrix of class "dgCMatrix"
#>          x1     x2     x3     x4 x3_normalized x4_normalized
#>  [1,] -0.11 -0.052  0.045 -0.320    0.21765254     0.1595718
#>  [2,]  0.28  .      0.790  .        1.05999612     0.7128002
#>  [3,] -1.50  1.400  1.300 -0.075    1.63663400     0.5831373
#>  [4,]  0.43  0.560 -0.710 -0.088   -0.63599767     0.5606624
#>  [5,]  1.00 -0.480  .     -0.071    0.16677272     0.5900526
#>  [6,] -0.71  2.000 -1.600  0.081   -1.64228732     0.8528361
#>  [7,]  0.52 -0.800 -0.610 -1.500   -0.52293142    -1.8804579
#>  [8,] -1.20  1.300  0.530 -0.590    0.76602386    -0.3072146
#>  [9,]  .    -0.140 -0.120 -0.160    0.03109322     0.4361860
#> [10,]  1.40  0.460 -1.100 -1.400   -1.07695605    -1.7075740