Transform Sparse Matrix columns

transform_cols() transforms specified matrix columns with a user-supplied function.

Usage

transform_cols(x, fun, ..., which, drop, name.sep)

# S3 method for dgCMatrix
transform_cols(x, fun, ..., which, drop = FALSE, name.sep = NULL)

# S3 method for matrix
transform_cols(x, fun, ..., which, drop = FALSE, name.sep = NULL)

Arguments

x

A matrix or dgCMatrix.

fun

A user-supplied function to apply to the specified columns.

...

Additional arguments to pass to fun.

which

A numeric vector indicating column indices or a character vector indicating column names.

drop

A logical value. If fun works with columnar matrices, then set drop = FALSE, otherwise if fun requires a numeric vector set drop = TRUE.

name.sep

A character vector of length 1 or nrow(x) that will be appended to existing column names to create new columns. Providing this argument ensures that the transformed columns will be appended as new matrix columns. If name.sep is NULL the specified columns will be transformed in-place. If name.sep is provided, but is faulty, the specified columns will be appended as new columns but will have un-changed column names.

Details

transform_cols() is an S3 generic with methods for:

dgCMatrix

matrix

Examples

x <- Matrix::rsparsematrix(10, 4, .9)
colnames(x) <- paste0("x", 1:4)
x
#> 10 x 4 sparse Matrix of class "dgCMatrix"
#>           x1    x2    x3     x4
#>  [1,] -2.600  0.38 -0.36  0.450
#>  [2,] -1.200 -1.00 -2.90  1.500
#>  [3,] -0.930  .     0.51 -1.300
#>  [4,]  0.650 -0.48 -0.45 -0.022
#>  [5,] -0.270 -0.47 -1.30 -0.160
#>  [6,] -0.025  1.20 -0.45  1.800
#>  [7,]  .     -0.15 -0.39  0.640
#>  [8,]  .      .     0.60  0.550
#>  [9,] -0.049 -0.62 -0.19  0.560
#> [10,]  1.500 -0.53 -1.70 -0.098

transform_cols(x, fun = function(i) i^2, which = paste0("x", 3:4))
#> 10 x 4 sparse Matrix of class "dgCMatrix"
#>           x1    x2     x3       x4
#>  [1,] -2.600  0.38 0.1296 0.202500
#>  [2,] -1.200 -1.00 8.4100 2.250000
#>  [3,] -0.930  .    0.2601 1.690000
#>  [4,]  0.650 -0.48 0.2025 0.000484
#>  [5,] -0.270 -0.47 1.6900 0.025600
#>  [6,] -0.025  1.20 0.2025 3.240000
#>  [7,]  .     -0.15 0.1521 0.409600
#>  [8,]  .      .    0.3600 0.302500
#>  [9,] -0.049 -0.62 0.0361 0.313600
#> [10,]  1.500 -0.53 2.8900 0.009604
transform_cols(x, fun = scale, which = 3:4, drop = TRUE)
#> 10 x 4 sparse Matrix of class "dgCMatrix"
#>           x1    x2         x3         x4
#>  [1,] -2.600  0.38  0.2883578  0.0662059
#>  [2,] -1.200 -1.00 -2.1288987  1.2647611
#>  [3,] -0.930  .     1.1163157 -1.9313860
#>  [4,]  0.650 -0.48  0.2027069 -0.4725732
#>  [5,] -0.270 -0.47 -0.6062175 -0.6300976
#>  [6,] -0.025  1.20  0.2027069  1.6072054
#>  [7,]  .     -0.15  0.2598075  0.2830873
#>  [8,]  .      .     1.2019665  0.1803540
#>  [9,] -0.049 -0.62  0.4501426  0.1917688
#> [10,]  1.500 -0.53 -0.9868878 -0.5593257
transform_cols(x, fun = scale, which = 3:4, name.sep = "normalized")
#> 10 x 6 sparse Matrix of class "dgCMatrix"
#>           x1    x2    x3     x4 x3_normalized x4_normalized
#>  [1,] -2.600  0.38 -0.36  0.450     0.2883578     0.0662059
#>  [2,] -1.200 -1.00 -2.90  1.500    -2.1288987     1.2647611
#>  [3,] -0.930  .     0.51 -1.300     1.1163157    -1.9313860
#>  [4,]  0.650 -0.48 -0.45 -0.022     0.2027069    -0.4725732
#>  [5,] -0.270 -0.47 -1.30 -0.160    -0.6062175    -0.6300976
#>  [6,] -0.025  1.20 -0.45  1.800     0.2027069     1.6072054
#>  [7,]  .     -0.15 -0.39  0.640     0.2598075     0.2830873
#>  [8,]  .      .     0.60  0.550     1.2019665     0.1803540
#>  [9,] -0.049 -0.62 -0.19  0.560     0.4501426     0.1917688
#> [10,]  1.500 -0.53 -1.70 -0.098    -0.9868878    -0.5593257